requirements.txt
build>=1.2,<2.0
jsonschema>=4.23,<5.0
mypy>=1.10,<2.0
pre-commit>=3.7,<5.0
pydantic>=2.6,<3.0
PyYAML>=6.0,<7.0
pytest>=8.0,<10.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
rich>=13.7,<16.0
ruff>=0.5,<1.0
twine>=6.0,<7.0
types-PyYAML>=6.0,<7.0
types-requests>=2.32,<3.0
typer>=0.12,<1.0
